Move copy paste code up to shared module
